R13/T07/feature/plans | refatoração e sidebar
- Refatorações efetuadas para maior legibilidade do código
  - PlanSelector agora é uma classe independente.
  - Removido bind repetido da store no módulo do plano.
  - Adicionados paddings para melhor experiência visual.
- Navegação pela sidebar implementada.
