+++ lib/app/global/widgets/side_bar/components_side_bar/buttons_side_bar.dart
@@ -1,3 +1,5 @@
+import 'package:pscomidas/app/modules/restaurant_home/pages/plans/restaurant_plan_page.dart';
+
 class ButtonSideBar {
   static final buttons = {
     'Produtos': {
@@ -8,6 +10,9 @@ class ButtonSideBar {
     },
     'Avaliações': {
       'navigator': 'avaliations',
+    },
+    'Planos': {
+      'navigator': RestaurantPlanPage.routeName,
     },
     'Sair': {
       'navigator': '/',
